refactor: rename schema to createSchema
