+++ apps/react-formbridge-docs/src/components/FeedbackPage.tsx
@@ -2,88 +2,17 @@
 
 import { type CSSProperties, useCallback, useMemo, useState } from 'react';
 
+import { type GlobaleDefaultsProps, useFormBridge } from '@runilib/react-formbridge';
 import {
-  field,
-  type GlobaleDefaultsProps,
-  type SchemaValues,
-  useFormBridge,
-} from '@runilib/react-formbridge';
-import {
-  FEEDBACK_TYPE_OPTIONS,
   type FeedbackSubmissionValues,
   getInitialFeedbackValues,
   normalizeFeedbackType,
   normalizeRelevantPage,
 } from '@/lib/feedback';
+import { FEEDBACK_SCHEMA, type FeedbackFormValues } from '@/lib/feedbackSchema';
 
 import Link from 'next/link';
 import styled, { css, type DefaultTheme, useTheme } from 'styled-components';
-
-const FEEDBACK_SCHEMA = {
-  feedbackType: field
-    .select('Feedback type')
-    .options(
-      FEEDBACK_TYPE_OPTIONS.map((option) => ({
-        label: option.label,
-        value: option.value,
-      })),
-    )
-    .defaultValue('general')
-    .required()
-    .hint('Pick the closest lane so the report starts in the right place.'),
-  subject: field
-    .text('Subject')
-    .required()
-    .min(4)
-    .placeholder('Short summary of your feedback')
-    .hint('A short title makes triage much faster.'),
-  area: field
-    .text('Area')
-    .placeholder('Documentation, validation adapters, wizard flows...')
-    .hint('Optional, but helpful when the topic is very specific.'),
-  relevantPage: field
-    .text('Relevant page')
-    .placeholder('/docs/useformbridge or full URL')
-    .hint('We prefill this when you open feedback from a docs page.'),
-  name: field.text('Name').placeholder('Optional'),
-  email: field
-    .email('Contact email')
-    .placeholder('name@company.com')
-    .hint('Optional. Leave it if you want a follow-up.'),
-  message: field
-    .textarea('Details')
-    .required()
-    .min(20)
-    .max(3000)
-    .placeholder('What did you try, what felt off, and what would make this better?')
-    .hint('The more concrete the context, the easier it is to act on.'),
-  expectedBehavior: field
-    .textarea('Expected behavior')
-    .visibleWhen('feedbackType', 'bug')
-    .requiredWhen('feedbackType', 'bug')
-    .clearOnHide()
-    .placeholder('What should have happened?')
-    .hint('Keep it short and observable.'),
-  actualBehavior: field
-    .textarea('Actual behavior')
-    .visibleWhen('feedbackType', 'bug')
-    .requiredWhen('feedbackType', 'bug')
-    .clearOnHide()
-    .placeholder('What actually happened?')
-    .hint('Include error copy, odd state, or incorrect result.'),
-  reproductionSteps: field
-    .textarea('Steps to reproduce')
-    .visibleWhen('feedbackType', 'bug')
-    .requiredWhen('feedbackType', 'bug')
-    .clearOnHide()
-    .placeholder('1. Go to...\n2. Click...\n3. Observe...')
-    .hint('A tiny repro is worth a lot.'),
-  contactConsent: field
-    .checkbox('You can contact me if follow-up details would help.')
-    .hint('Useful if you left a contact email above.'),
-};
-
-type FeedbackFormValues = SchemaValues<typeof FEEDBACK_SCHEMA>;
 
 function getFeedbackFieldStyles(theme: DefaultTheme): GlobaleDefaultsProps {
   const errorColor = theme.mode === 'dark' ? '#ff9fb0' : '#d1435b';
